Move general REST API classes to com.yahoo.restapi
