+++ container-core/abi-spec.json
@@ -854,6 +854,57 @@
     ],
     "fields": []
   },
+  "com.yahoo.restapi.ErrorResponse$errorCodes": {
+    "superClass": "java.lang.Enum",
+    "interfaces": [],
+    "attributes": [
+      "public",
+      "final",
+      "enum"
+    ],
+    "methods": [
+      "public static com.yahoo.restapi.ErrorResponse$errorCodes[] values()",
+      "public static com.yahoo.restapi.ErrorResponse$errorCodes valueOf(java.lang.String)"
+    ],
+    "fields": [
+      "public static final enum com.yahoo.restapi.ErrorResponse$errorCodes NOT_FOUND",
+      "public static final enum com.yahoo.restapi.ErrorResponse$errorCodes BAD_REQUEST",
+      "public static final enum com.yahoo.restapi.ErrorResponse$errorCodes FORBIDDEN",
+      "public static final enum com.yahoo.restapi.ErrorResponse$errorCodes METHOD_NOT_ALLOWED",
+      "public static final enum com.yahoo.restapi.ErrorResponse$errorCodes INTERNAL_SERVER_ERROR",
+      "public static final enum com.yahoo.restapi.ErrorResponse$errorCodes UNAUTHORIZED"
+    ]
+  },
+  "com.yahoo.restapi.ErrorResponse": {
+    "superClass": "com.yahoo.restapi.SlimeJsonResponse",
+    "interfaces": [],
+    "attributes": [
+      "public"
+    ],
+    "methods": [
+      "public void <init>(int, java.lang.String, java.lang.String)",
+      "public static com.yahoo.restapi.ErrorResponse notFoundError(java.lang.String)",
+      "public static com.yahoo.restapi.ErrorResponse internalServerError(java.lang.String)",
+      "public static com.yahoo.restapi.ErrorResponse badRequest(java.lang.String)",
+      "public static com.yahoo.restapi.ErrorResponse forbidden(java.lang.String)",
+      "public static com.yahoo.restapi.ErrorResponse unauthorized(java.lang.String)",
+      "public static com.yahoo.restapi.ErrorResponse methodNotAllowed(java.lang.String)"
+    ],
+    "fields": []
+  },
+  "com.yahoo.restapi.MessageResponse": {
+    "superClass": "com.yahoo.container.jdisc.HttpResponse",
+    "interfaces": [],
+    "attributes": [
+      "public"
+    ],
+    "methods": [
+      "public void <init>(java.lang.String)",
+      "public void render(java.io.OutputStream)",
+      "public java.lang.String getContentType()"
+    ],
+    "fields": []
+  },
   "com.yahoo.restapi.Path": {
     "superClass": "java.lang.Object",
     "interfaces": [],
@@ -869,6 +920,63 @@
       "public java.lang.String get(java.lang.String)",
       "public java.lang.String getRest()",
       "public java.lang.String asString()",
+      "public java.lang.String toString()"
+    ],
+    "fields": []
+  },
+  "com.yahoo.restapi.ResourceResponse": {
+    "superClass": "com.yahoo.container.jdisc.HttpResponse",
+    "interfaces": [],
+    "attributes": [
+      "public"
+    ],
+    "methods": [
+      "public varargs void <init>(com.yahoo.container.jdisc.HttpRequest, java.lang.String[])",
+      "public void render(java.io.OutputStream)",
+      "public java.lang.String getContentType()"
+    ],
+    "fields": []
+  },
+  "com.yahoo.restapi.SlimeJsonResponse": {
+    "superClass": "com.yahoo.container.jdisc.HttpResponse",
+    "interfaces": [],
+    "attributes": [
+      "public"
+    ],
+    "methods": [
+      "public void <init>(com.yahoo.slime.Slime)",
+      "public void <init>(int, com.yahoo.slime.Slime)",
+      "public void render(java.io.OutputStream)",
+      "public java.lang.String getContentType()"
+    ],
+    "fields": []
+  },
+  "com.yahoo.restapi.StringResponse": {
+    "superClass": "com.yahoo.container.jdisc.HttpResponse",
+    "interfaces": [],
+    "attributes": [
+      "public"
+    ],
+    "methods": [
+      "public void <init>(java.lang.String)",
+      "public void render(java.io.OutputStream)"
+    ],
+    "fields": []
+  },
+  "com.yahoo.restapi.Uri": {
+    "superClass": "java.lang.Object",
+    "interfaces": [],
+    "attributes": [
+      "public"
+    ],
+    "methods": [
+      "public void <init>(java.net.URI)",
+      "public void <init>(java.lang.String)",
+      "public com.yahoo.restapi.Uri append(java.lang.String)",
+      "public com.yahoo.restapi.Uri withoutParameters()",
+      "public com.yahoo.restapi.Uri withPath(java.lang.String)",
+      "public com.yahoo.restapi.Uri withTrailingSlash()",
+      "public java.net.URI toURI()",
       "public java.lang.String toString()"
     ],
     "fields": []
