+++ container-core/src/main/java/com/yahoo/restapi/ErrorResponse.java
@@ -1,5 +1,5 @@
 // Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.vespa.configserver.flags.http;
+package com.yahoo.restapi;
 
 import com.yahoo.slime.Cursor;
 import com.yahoo.slime.Slime;
