+++ container-core/src/main/java/com/yahoo/restapi/MessageResponse.java
@@ -1,5 +1,5 @@
 // Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.vespa.hosted.controller.restapi;
+package com.yahoo.restapi;
 
 import com.yahoo.container.jdisc.HttpResponse;
 import com.yahoo.slime.JsonFormat;
